move formatter files to its own folder
