+++ src/ensembl/src/shared/helpers/formatters/numberFormater.test.ts
