+++ src/ensembl/src/shared/helpers/formatters/numberFormatter.ts
