+++ src/ensembl/src/shared/helpers/formatters/strandFormatter.ts
